+++ Source/Krypton Components/Krypton.Toolkit.Utilities/Globals/GlobalDeclarations.cs
@@ -11,6 +11,7 @@
 
 global using System;
 global using System.Collections;
+global using System.Collections.Concurrent;
 global using System.Collections.Generic;
 global using System.Collections.ObjectModel;
 global using System.Collections.Specialized;